Update architecture label
I thought it was just qemu supported stuff, but there are docker images to download &c. so trying one of those. The Raspberry Pi OS packages call it armhf, but I think that can apply to armv7 too (Pi Zero is armv6) so I don't know if it will mean the same thing.